% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{decisionTree}
\alias{decisionTree}
\title{decisiontree}
\usage{
decisionTree(
  data = data,
  formula = formula,
  engine = "glm",
  mode = "classification"
)
}
\arguments{
\item{data}{분석에 사용할 변수들이 포함된 데이터}

\item{engine}{engine}

\item{mode}{mode}
}
\description{
decisiontree
}
\details{
의사결정나무 알고리즘 함수. 의사 결정 규칙 (Decision rule)을 나무 형태로 분류해 나가는 분석 기법을 말합니다.
}
