% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/modeling.R
=======
% Please edit documentation in R/fitting.R
>>>>>>> 8906c592ae2fe4dc324f2b903cdd54cb904a5190
\name{fitBestModel}
\alias{fitBestModel}
\title{fitting in best model}
\usage{
fitBestModel(
  gridSearchResult,
  metric,
  model,
  formula,
  trainingData,
  splitedData
)
}
\arguments{
\item{gridSearchResult}{gridSearchResult}

\item{metric}{metric}

\item{model}{model}

\item{formula}{formula}

\item{trainingData}{trainingData}

\item{splitedData}{splitedData}
}
\description{
fitting in best model
}
\details{
fitting in best model
}
