% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/modeling.R
=======
% Please edit documentation in R/fitting.R
>>>>>>> 8906c592ae2fe4dc324f2b903cdd54cb904a5190
\name{gridSerachCV}
\alias{gridSerachCV}
\title{Grid Search with cross validation}
\usage{
<<<<<<< HEAD
gridSerachCV(rec, model, v = 5, data, parameterGrid)
=======
gridSerachCV(rec, model, v = "5", data, parameterGrid)
>>>>>>> 8906c592ae2fe4dc324f2b903cdd54cb904a5190
}
\arguments{
\item{rec}{rec}

\item{model}{model}

\item{v}{v-fold CV}

\item{data}{data}

\item{parameter_grid}{parameter_grid}
}
\description{
Grid Search with cross validation
}
\details{
Grid Search with cross validation
}
