% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{knn}
\alias{knn}
\title{K-NN}
\usage{
knn(data = data, formula = formula, engine = "kknn", mode = "classification")
}
\arguments{
\item{data}{분석에 사용할 변수들이 포함된 데이터}

\item{engine}{engine}

\item{mode}{mode}
}
\description{
K-NN
}
\details{
K-NN 알고리즘 함수.
데이터로부터 거리가 가까운 K개의 다른 데이터의 레이블을 참조하여 분류하는 알고리즘
}
