% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{lightGbm_phi}
\alias{lightGbm_phi}
\title{Light GBM}
\usage{
lightGbm_phi(engine = "lightgbm", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Light GBM
}
\details{
Light GBM
install treesnip package by: remotes::install_github("curso-r/treesnip")
hyperparameters: mtry, min_n, tree_depth, loss_reduction, learn_rate, sample_size
}
