% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R, R/supervisedLearning.R
\name{naiveBayes_phi}
\alias{naiveBayes_phi}
\title{Naive Bayes}
\usage{
naiveBayes_phi(engine = "klaR", mode = "classification")

naiveBayes_phi(engine = "klaR", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Naive Bayes

Naive Bayes
}
\details{
Naive Bayes
hyperparameters: smoothness, Laplace

Naive Bayes
hyperparameters: smoothness, Laplace
}
