% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{randomForest}
\alias{randomForest}
\title{random Forest}
\usage{
randomForest(
  data = data,
  formula = formula,
  engine = "randomForest",
  mode = "classification"
)
}
\arguments{
\item{data}{분석에 사용할 변수들이 포함된 데이터}

\item{engine}{engine}

\item{mode}{mode}
}
\description{
random Forest
}
\details{
랜덤 포레스트 알고리즘 함수. 여러개의 Decision Tree를 형성.
새로운 데이터 포인트를 각 트리에 동시에 통과 시켜 각 트리가 분류한 결과에서 투표를 실시하여
가장 많이 득표한 결과를 최종 분류 결과로 선택
}
