% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{svm}
\alias{svm}
\title{svm}
\usage{
svm(
  data = data,
  formula = formula,
  engine = "LiblineaR",
  mode = "classification"
)
}
\arguments{
\item{data}{분석에 사용할 변수들이 포함된 데이터}

\item{engine}{engine}

\item{mode}{mode}
}
\description{
svm
}
\details{
support vector machines 알고리즘 함수.
support vector machines 기계 학습의 분야 중 하나로 패턴 인식, 자료 분석을 위한 지도 학습 모델
}
