% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{xgboost}
\alias{xgboost}
\title{xgboost}
\usage{
xgboost(
  data = data,
  formula = formula,
  engine = "xgboost",
  mode = "classification"
)
}
\arguments{
\item{data}{분석에 사용할 변수들이 포함된 데이터}

\item{engine}{engine}

\item{mode}{mode}
}
\description{
xgboost
}
\details{
xgboost 알고리즘 함수.
xgboost
}
