% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{xgboost_phi}
\alias{xgboost_phi}
\title{xgboost}
\usage{
xgboost_phi(engine = "xgboost", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}

\item{data}{분석에 사용할 변수들이 포함된 데이터}
}
\description{
xgboost
}
\details{
xgboost 알고리즘 함수.
xgboost는 Extreme Gradient Boosting의 약자.
Gradient Boost 알고리즘을 병렬 학습이 지원되도록 구현한 라이브러리
}
