% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/supervisedLearning.R
=======
% Please edit documentation in R/algorithms.R
>>>>>>> 8906c592ae2fe4dc324f2b903cdd54cb904a5190
\name{lightGbm_phi}
\alias{lightGbm_phi}
\title{Light GBM}
\usage{
lightGbm_phi(engine = "lightgbm", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Light GBM
}
\details{
Light GBM
hyperparameters: mtry, min_n, tree_depth, loss_reduction, learn_rate, sample_size
}
