% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{lightGbm_phi}
\alias{lightGbm_phi}
\title{Light GBM}
\usage{
lightGbm_phi(
  mtry = 5,
  trees = 1000,
  min_n = tune(),
  tree_depth = tune(),
  loss_reduction = tune(),
  learn_rate = tune(),
  sample_size = 0.7,
  engine = "lightgbm",
  mode = "classification"
)
}
\arguments{
\item{mtry}{mtry}

\item{trees}{trees}

\item{min_n}{min_n}

\item{tree_depth}{tree_depth}

\item{loss_reduction}{loss_reduction}

\item{learn_rate}{learn_rate}

\item{sample_size}{sample_size}

\item{engine}{engine}

\item{mode}{mode}
}
\description{
Light GBM
}
\details{
Light GBM
}
