% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/supervisedLearning.R
=======
% Please edit documentation in R/algorithms.R
>>>>>>> 8906c592ae2fe4dc324f2b903cdd54cb904a5190
\name{naiveBayes_phi}
\alias{naiveBayes_phi}
\title{Naive Bayes}
\usage{
naiveBayes_phi(engine = "klaR", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Naive Bayes
}
\details{
Naive Bayes
hyperparameters: smoothness, Laplace
}
